Moving UI elements to pixel package
