+++ ime/app/src/main/res/layout/quick_text_popup_autorowkeyboard_view.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<com.anysoftkeyboard.ui.ScrollViewWithDisable xmlns:android="http://schemas.android.com/apk/res/android"
+<net.evendanan.pixel.ScrollViewWithDisable xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/scroll_root_for_quick_test_keyboard"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -11,4 +11,4 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:clipToPadding="false"/>
-</com.anysoftkeyboard.ui.ScrollViewWithDisable>
+</net.evendanan.pixel.ScrollViewWithDisable>
