+++ ime/app/src/main/res/layout/quick_text_popup_root_view.xml
@@ -18,12 +18,12 @@
             android:layout_height="wrap_content"
             android:paddingTop="4dp" android:paddingBottom="4dp"/>
 
-        <com.anysoftkeyboard.ui.ViewPagerWithDisable
+        <net.evendanan.pixel.ViewPagerWithDisable
             android:id="@+id/quick_text_keyboards_pager"
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
 
-        </com.anysoftkeyboard.ui.ViewPagerWithDisable>
+        </net.evendanan.pixel.ViewPagerWithDisable>
 
     </LinearLayout>
