+++ ime/app/src/test/java/com/anysoftkeyboard/quicktextkeys/ui/QuickKeysKeyboardPagerAdapterTest.java
@@ -10,11 +10,11 @@ import com.anysoftkeyboard.keyboards.views.AnyKeyboardViewWithMiniKeyboard;
 import com.anysoftkeyboard.keyboards.views.OnKeyboardActionListener;
 import com.anysoftkeyboard.keyboards.views.QuickKeysKeyboardView;
 import com.anysoftkeyboard.quicktextkeys.QuickTextKey;
-import com.anysoftkeyboard.ui.ScrollViewWithDisable;
-import com.anysoftkeyboard.ui.ViewPagerWithDisable;
 import com.menny.android.anysoftkeyboard.AnyApplication;
 import com.menny.android.anysoftkeyboard.R;
 import java.util.List;
+import net.evendanan.pixel.ScrollViewWithDisable;
+import net.evendanan.pixel.ViewPagerWithDisable;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
