+++ ime/app/src/test/java/com/anysoftkeyboard/quicktextkeys/ui/QuickTextPagerViewTest.java
@@ -15,10 +15,10 @@ import com.anysoftkeyboard.quicktextkeys.QuickTextKeyFactory;
 import com.anysoftkeyboard.remote.MediaType;
 import com.anysoftkeyboard.rx.TestRxSchedulers;
 import com.anysoftkeyboard.theme.KeyboardTheme;
-import com.anysoftkeyboard.ui.ViewPagerWithDisable;
 import com.menny.android.anysoftkeyboard.AnyApplication;
 import com.menny.android.anysoftkeyboard.R;
 import java.util.Collections;
+import net.evendanan.pixel.ViewPagerWithDisable;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
