+++ ime/app/src/testAllAddOns/java/com/anysoftkeyboard/keyboards/LoadedKeyboardsVerificationTest.java
@@ -121,7 +121,8 @@ public class LoadedKeyboardsVerificationTest {
                   + " char "
                   + ((char) key.getPrimaryCode());
           Assert.assertTrue(keyId, key.x >= 0);
-          Assert.assertTrue(keyId, Keyboard.Key.getEndX(key) <= sTestKeyboardDimens.getKeyboardMaxWidth());
+          Assert.assertTrue(
+              keyId, Keyboard.Key.getEndX(key) <= sTestKeyboardDimens.getKeyboardMaxWidth());
           Assert.assertTrue(keyId, key.y >= 0);
         }
       }
