+++ ime/pixel/src/main/java/net/evendanan/pixel/ViewPagerWithDisable.java
@@ -1,4 +1,4 @@
-package com.anysoftkeyboard.ui;
+package net.evendanan.pixel;
 
 import android.content.Context;
 import android.util.AttributeSet;
